Use float epsilon for float maths
Make an additional epsilon scaling function for float epsilon,
and convert all epsilon reference when doing CoordType operations
over to CoordType. This should still allow CoordType to be
switch at will to double.

I spent a long time looking into whether there is a reason to use
double for scaling epsilon while using float for math, and there
really doesn't seem to be. Using float (as CoordType) everywhere
gets rid of expensive conversions thanks to the availability
of float specific cmath function in the C++ library headers.

+++ wfmath/quaternion.cpp
@@ -61,26 +61,26 @@ Quaternion::Quaternion (CoordType w_in,
 // correspond to the same rotation matrix. We consider the form
 // of the quaternion with w > 0 canonical.
 
-bool Quaternion::isEqualTo(const Quaternion &q, double epsilon) const
+bool Quaternion::isEqualTo(const Quaternion &q, float epsilon) const
 {
   // Since the sum of squares is 1, the magnitude of the largest
   // element must be between 1 and 0.5, so we don't need to scale epsilon.
 
   assert(epsilon > 0);
 
-  if(fabs(m_w - q.m_w) <= epsilon) {
+  if(std::fabs(m_w - q.m_w) <= epsilon) {
     int i;
     for(i = 0; i < 3; ++i)
-      if(fabs(m_vec[i] - q.m_vec[i]) > epsilon)
+      if(std::fabs(m_vec[i] - q.m_vec[i]) > epsilon)
         break; // try again with swapped signs
     if(i == 3) // got through the loop
       return true;
   }
 
   // This makes q == -q true
-  if(fabs(m_w + q.m_w) <= epsilon) {
+  if(std::fabs(m_w + q.m_w) <= epsilon) {
     for(int i = 0; i < 3; ++i)
-      if(fabs(m_vec[i] + q.m_vec[i]) > epsilon)
+      if(std::fabs(m_vec[i] + q.m_vec[i]) > epsilon)
         return false;
     return true;
   }
